CloudFunctionsCode/requirements.txt
mysql-connector-python
google-cloud-pubsub

mysql-connector-python